requirements.txt
Django==1.8.2
beautifulsoup4==4.4.0
feedparser==5.2.0
python-dateutil==2.4.2
requests==2.7.0
six==1.9.0
